cookify/requirements.txt
# Core dependencies
numpy>=1.24.0  # Use newer version compatible with Python 3.12
pandas>=1.5.0  # Use newer version compatible with Python 3.12
matplotlib>=3.7.0  # Use newer version compatible with Python 3.12
tqdm>=4.65.0

# Video processing
opencv-python>=4.5.3
ffmpeg-python>=0.2.0
scenedetect==0.6.1  # Pin to a specific version to avoid compatibility issues

# Machine learning and computer vision
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0  # For YOLOv8
# mmaction2>=1.0.0  # Commented out as it may have installation issues

# OCR and text recognition
# easyocr>=1.5.0  # Commented out due to Python 3.12 compatibility issues
pytesseract>=0.3.8
Pillow>=8.3.1

# Audio processing and speech recognition
# openai-whisper>=1.0.0  # Commented out due to Python 3.12 compatibility issues
librosa>=0.8.1      # Audio processing

# NLP
spacy>=3.1.0
nltk>=3.6.2
transformers>=4.9.1

# Utilities
jsonschema>=3.2.0
PyYAML>=6.0
requests>=2.26.0

# Web interface
Flask>=2.0.1
Werkzeug>=2.0.1
Jinja2>=3.0.1
itsdangerous>=2.0.1
click>=8.0.1

# Note: Some dependencies have been commented out due to compatibility issues with Python 3.12.
# If you're using Python 3.12, you may need to downgrade to Python 3.11 or 3.10 to use all features.
# Alternatively, you can try installing these packages manually with pip install --no-deps.
